Scene creation has been cleaned up a bit
