+++ .gitignore
@@ -1,6 +1,7 @@
 # Mine
 MVC/db
 minio/data
+Pong copy
 
 # Keys
 private.key
