+++ client/src/scene/scene.type.ts
@@ -1,0 +1,13 @@
+export interface PaddleConstructor {
+	paddleName: string,
+	dimensions: PaddleDimensions,
+	positionX: number,
+	positionY: number,
+};
+
+interface PaddleDimensions {
+	size: number,
+	width: number,
+	height: number,
+	depth: number,
+};
